Move the tests for an admin client into a test helper.  Reuse this in all 3 places that we return the admin API client
